+++ proxy/proxy.py
@@ -12,13 +12,12 @@ import contextlib
 import os
 import sys
 import time
-import unittest
 import logging
 
 from types import TracebackType
 from typing import List, Optional, Generator, Any, Type
 
-from .common.utils import bytes_, get_available_port, new_socket_connection
+from .common.utils import bytes_
 from .common.flags import Flags
 from .core.acceptor import AcceptorPool
 from .http.handler import HttpProtocolHandler
@@ -58,63 +57,6 @@ class Proxy:
         assert self.acceptors
         self.acceptors.shutdown()
         self.delete_pid_file()
-
-
-class TestCase(unittest.TestCase):
-    """Base TestCase class that automatically setup and teardown proxy.py."""
-
-    DEFAULT_PROXY_PY_STARTUP_FLAGS = [
-        '--num-workers', '1',
-        '--threadless',
-    ]
-
-    PROXY: Optional[Proxy] = None
-    PROXY_PORT: int = 8899
-    INPUT_ARGS: Optional[List[str]] = None
-    ENABLE_VCR: bool = False
-
-    @classmethod
-    def setUpClass(cls) -> None:
-        cls.PROXY_PORT = get_available_port()
-        cls.INPUT_ARGS = getattr(cls, 'PROXY_PY_STARTUP_FLAGS') \
-            if hasattr(cls, 'PROXY_PY_STARTUP_FLAGS') \
-            else cls.DEFAULT_PROXY_PY_STARTUP_FLAGS
-        cls.INPUT_ARGS.append('--port')
-        cls.INPUT_ARGS.append(str(cls.PROXY_PORT))
-        cls.PROXY = Proxy(input_args=cls.INPUT_ARGS)
-        cls.PROXY.__enter__()
-        cls.wait_for_server(cls.PROXY_PORT)
-
-    @staticmethod
-    def wait_for_server(proxy_port: int) -> None:
-        """Wait for proxy.py server to come up."""
-        while True:
-            try:
-                conn = new_socket_connection(
-                    ('localhost', proxy_port))
-                conn.close()
-                break
-            except ConnectionRefusedError:
-                time.sleep(0.1)
-
-    @classmethod
-    def tearDownClass(cls) -> None:
-        assert cls.PROXY
-        cls.PROXY.__exit__(None, None, None)
-        cls.PROXY = None
-        cls.PROXY_PORT = 8899
-        cls.INPUT_ARGS = None
-
-    @contextlib.contextmanager
-    def vcr(self) -> Generator[None, None, None]:
-        self.ENABLE_VCR = True
-        try:
-            yield
-        finally:
-            self.ENABLE_VCR = False
-
-    def run(self, result: Optional[unittest.TestResult] = None) -> Any:
-        super().run(result)
 
 
 @contextlib.contextmanager
